cart

+++ CA1Assignment/src/main/java/models/AdminService.java
@@ -20,7 +20,7 @@ public class AdminService {
 				String dbName = rs.getString("tour_name");
 				String dbBriefDescription = rs.getString("brief_description");
 				String dbFullDescription = rs.getString("detail_description");
-				int dbPrice = rs.getInt("price");
+				double dbPrice = rs.getInt("price");
 				int dbAvailableSlots = rs.getInt("available_slots");
 				int dbCategoryID = rs.getInt("tour_category_id");
 				String dbImage = rs.getString("image_location");
@@ -57,7 +57,7 @@ public class AdminService {
 			}
 		return numRowsAffected;
 	}
-	public int adminUpdate(int id, String name, String briefDescription, String fullDescription, int price, int slots, int catID, String image) {
+	public int adminUpdate(int id, String name, String briefDescription, String fullDescription, double price, int slots, int catID, String image) {
 		int numRowsAffected = 0;
 		try {
 			Class.forName("com.mysql.jdbc.Driver");
@@ -69,7 +69,7 @@ public class AdminService {
 			ps.setString(1,name);
 			ps.setString(2,briefDescription);
 			ps.setString(3,fullDescription);
-			ps.setInt(4,price);
+			ps.setDouble(4,price);
 			ps.setInt(5,slots);
 			ps.setInt(6,catID);
 			ps.setString(7,image);
@@ -81,7 +81,7 @@ public class AdminService {
 			}
 		return numRowsAffected;
 	}
-	public int adminInsert(String name, String briefDescription, String fullDescription, int price, int slots, int catID, String image) {
+	public int adminInsert(String name, String briefDescription, String fullDescription, double price, int slots, int catID, String image) {
 		int numRowsAffected = 0;
 		try {
 			Class.forName("com.mysql.jdbc.Driver");
@@ -93,7 +93,7 @@ public class AdminService {
 			ps.setString(1, name);
 			ps.setString(2, briefDescription);
 			ps.setString(3, fullDescription);
-			ps.setInt(4, price);
+			ps.setDouble(4, price);
 			ps.setInt(5, slots);
 			ps.setInt(6, catID);
 			ps.setString(7, image);
